+++ sanakirja/cli.py
@@ -7,9 +7,9 @@ from typing import Set, Union
 
 # All possible fields returned by the `search` method of the `Sanakirja` object
 AVAILABLE_FIELDS = {
-    "id", "source_language", "target_language", "word", "transliteration", "additional_source_languages",
-    "relations", "similar_words", "alternative_spellings", "multiple_spellings", "synonyms",
-    "pronunciations", "abbreviations", "inflections", "definitions", "examples", "categories", "translations"
+    "id", "source_language", "target_language", "word", "transliteration", "gender", "additional_source_languages",
+    "relations", "did_you_mean", "suggestions", "found_examples", "alternative_spellings", "multiple_spellings", 
+    "synonyms", "pronunciations", "abbreviations", "inflections", "definitions", "examples", "categories", "translations"
 }
 
 def _parse_lang(lang: str) -> Union[int, str]:
@@ -61,34 +61,28 @@ def main() -> None:
     args = parser.parse_args()
     sk = sanakirja.Sanakirja()
 
-    try:
-        result = sk.search(args.word, args.source, args.target)
-        omitted = _resolve_show_fields(args.show)
-        output = {k: v for k, v in result.items() if k not in omitted}
-        if not args.output: return print(json.dumps(output, indent=args.pretty, ensure_ascii=False))
+    result = sk.search(args.word, args.source, args.target)
+    omitted = _resolve_show_fields(args.show)
+    output = {k: v for k, v in result.items() if k not in omitted}
+    if not args.output: return print(json.dumps(output, indent=args.pretty, ensure_ascii=False))
 
-        out_path = Path(args.output).expanduser()
+    out_path = Path(args.output).expanduser()
 
-        if out_path.exists() and out_path.is_dir():
-            return print(f"\033[31mERROR\033[0m: The path '{out_path}' is a directory.")
-        elif not out_path.parent.exists():
-            return print(f"\033[31mERROR\033[0m: The directory '{out_path.parent}' does not exist.")
-        elif not os.access(out_path.parent, os.W_OK):
-            return print(f"\033[31mERROR\033[0m: Write access to '{out_path.parent}' denied.")
+    if out_path.exists() and out_path.is_dir():
+        return print(f"\033[31mERROR\033[0m: The path '{out_path}' is a directory.")
+    elif not out_path.parent.exists():
+        return print(f"\033[31mERROR\033[0m: The directory '{out_path.parent}' does not exist.")
+    elif not os.access(out_path.parent, os.W_OK):
+        return print(f"\033[31mERROR\033[0m: Write access to '{out_path.parent}' denied.")
 
-        if os.path.exists(args.output):
-            confirm = input(f"The file '{args.output}' already exists. Overwrite? (Y/n): ").strip().lower()
-            if confirm not in ["y", "yes", ""]: return print("Aborted saving to file.")
+    if os.path.exists(args.output):
+        confirm = input(f"The file '{args.output}' already exists. Overwrite? (Y/n): ").strip().lower()
+        if confirm not in ["y", "yes", ""]: return print("Aborted saving to file.")
 
-        try:
-            with open(args.output, "w", encoding="utf-8") as f:
-                json.dump(output, f, ensure_ascii=False, indent=args.pretty)
-        except Exception as e:
-            return print("\033[31mERROR\033[0m:", e)
+    with open(args.output, "w", encoding="utf-8") as f:
+        json.dump(output, f, ensure_ascii=False, indent=args.pretty)
 
-        print(f"Saved output to '{args.output}'")
-    except Exception as e:
-        print("An error occurred:", e)
+    print(f"Saved output to '{args.output}'")
 
 if __name__ == "__main__":
     main()